@@ -1,69 +1,48 @@
 {
-  "name": "buymylooks-app",
+  "name": "client",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": " ng serve  --port 4300 --proxy-config proxy.config.json --host 0.0.0.0",
-    "build": "ng build --prod",
+    "start": "ng serve --port 4300 --proxy-config proxy.config.json ",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "deploy": "  ng build --prod && cp -r ./dist/buymylooks-app/* /var/www/html"
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.3",
-    "@angular/animations": "^6.0.3",
-    "@angular/cdk": "^6.2.0",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/flex-layout": "^6.0.0-beta.17",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/material": "^6.4.2",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/pwa": "^0.6.8",
-    "@angular/router": "^6.0.3",
-    "@angular/service-worker": "^6.0.3",
-    "@swimlane/ngx-charts": "^9.0.0",
-    "@types/chart.js": "^2.7.30",
-    "agm-direction": "^0.6.0",
-    "angular-bootstrap-md": "^6.2.2",
-    "angular5-csv": "^0.2.10",
-    "bootstrap": "^4.1.3",
-    "chart.js": "^2.5.0",
+    "@angular/animations": "^6.1.0",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
     "core-js": "^2.5.4",
-    "font-awesome": "^4.7.0",
-    "hammerjs": "^2.0.8",
-    "http-server": "^0.11.1",
-    "jquery": "^3.3.1",
-    "ngx-cookie-service": "^1.0.10",
-    "ngx-drag-scroll": "^2.0.0-beta.4",
-    "rxjs": "^6.0.0",
-    "zone.js": "^0.8.26"
+    "rxjs": "~6.2.0",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^6.0.3",
-    "@angular-devkit/build-angular": "~0.6.8",
-    "typescript": "~2.7.2",
-    "@angular/cli": "~6.0.8",
-    "@angular/language-service": "^6.0.3",
-    "@types/jasmine": "~2.8.6",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "~6.2.3",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "~4.3.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
+    "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "perfect-scrollbar": "^1.1.0"
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~2.9.2"
   }
 }
