@@ -1,4 +1,6 @@
-const { promisify } = require('util');
+const {
+    promisify
+} = require('util');
 const crypto = require('crypto');
 const nodemailer = require('nodemailer');
 const passport = require('passport');
@@ -16,8 +18,7 @@ const randomBytesAsync = promisify(crypto.randomBytes);
 exports.getLogin = (req, res) => {
     if (req.user) {
         return res.redirect('http://localhost:4200/Dashboard');
-    }
-    else {
+    } else {
         return res.redirect('http://localhost:4200/Login');
 
     }
@@ -30,7 +31,9 @@ exports.getLogin = (req, res) => {
 exports.postLogin = (req, res, next) => {
     req.assert('email', 'Email is not valid').isEmail();
     req.assert('password', 'Password cannot be blank').notEmpty();
-    req.sanitize('email').normalizeEmail({ gmail_remove_dots: false });
+    req.sanitize('email').normalizeEmail({
+        gmail_remove_dots: false
+    });
 
     const errors = req.validationErrors();
 
@@ -51,7 +54,9 @@ exports.postLogin = (req, res, next) => {
             if (err) {
                 return next(err);
             }
-            req.flash('success', { msg: 'Success! You are logged in.' });
+            req.flash('success', {
+                msg: 'Success! You are logged in.'
+            });
             res.redirect(req.session.returnTo || '/');
         });
     })(req, res, next);
@@ -91,7 +96,9 @@ exports.postSignup = (req, res, next) => {
     req.assert('email', 'Email is not valid').isEmail();
     req.assert('password', 'Password must be at least 4 characters long').len(4);
     req.assert('confirmPassword', 'Passwords do not match').equals(req.body.password);
-    req.sanitize('email').normalizeEmail({ gmail_remove_dots: false });
+    req.sanitize('email').normalizeEmail({
+        gmail_remove_dots: false
+    });
 
     const errors = req.validationErrors();
 
@@ -105,12 +112,16 @@ exports.postSignup = (req, res, next) => {
         password: req.body.password
     });
 
-    User.findOne({ email: req.body.email }, (err, existingUser) => {
+    User.findOne({
+        email: req.body.email
+    }, (err, existingUser) => {
         if (err) {
             return next(err);
         }
         if (existingUser) {
-            req.flash('errors', { msg: 'Account with that email address already exists.' });
+            req.flash('errors', {
+                msg: 'Account with that email address already exists.'
+            });
             return res.redirect('/signup');
         }
         user.save((err) => {
@@ -150,7 +161,9 @@ exports.getAccountDetail = (req, res) => {
  */
 exports.postUpdateProfile = (req, res, next) => {
     req.assert('email', 'Please enter a valid email address.').isEmail();
-    req.sanitize('email').normalizeEmail({ gmail_remove_dots: false });
+    req.sanitize('email').normalizeEmail({
+        gmail_remove_dots: false
+    });
 
     const errors = req.validationErrors();
 
@@ -171,12 +184,16 @@ exports.postUpdateProfile = (req, res, next) => {
         user.save((err) => {
             if (err) {
                 if (err.code === 11000) {
-                    req.flash('errors', { msg: 'The email address you have entered is already associated with an account.' });
+                    req.flash('errors', {
+                        msg: 'The email address you have entered is already associated with an account.'
+                    });
                     return res.redirect('/account');
                 }
                 return next(err);
             }
-            req.flash('success', { msg: 'Profile information has been updated.' });
+            req.flash('success', {
+                msg: 'Profile information has been updated.'
+            });
             res.redirect('/account');
         });
     });
@@ -206,7 +223,9 @@ exports.postUpdatePassword = (req, res, next) => {
             if (err) {
                 return next(err);
             }
-            req.flash('success', { msg: 'Password has been changed.' });
+            req.flash('success', {
+                msg: 'Password has been changed.'
+            });
             res.redirect('/account');
         });
     });
@@ -217,12 +236,16 @@ exports.postUpdatePassword = (req, res, next) => {
  * Delete user account.
  */
 exports.postDeleteAccount = (req, res, next) => {
-    User.remove({ _id: req.user.id }, (err) => {
+    User.remove({
+        _id: req.user.id
+    }, (err) => {
         if (err) {
             return next(err);
         }
         req.logout();
-        req.flash('info', { msg: 'Your account has been deleted.' });
+        req.flash('info', {
+            msg: 'Your account has been deleted.'
+        });
         res.redirect('/');
     });
 };
@@ -232,7 +255,8 @@ exports.postDeleteAccount = (req, res, next) => {
  * Post User Photos.
  */
 exports.importUserPhotos = (req, res, next) => {
-    let request_object = req.body, failed = passed = 0;
+    let request_object = req.body,
+        failed = passed = 0;
     for (const photos of req.body) {
         if (request_object instanceof Array) {
             for (let i = 0; i < request_object.length; i++) {
@@ -246,11 +270,19 @@ exports.importUserPhotos = (req, res, next) => {
                     permalink: request_object[i].permalink,
                     username: request_object[i].username,
                 });
-                InstagramPhotos.findOne({ _id: request_object[i]._id }).then((existingPhoto) => {
+                InstagramPhotos.findOne({
+                    _id: request_object[i]._id
+                }).then((existingPhoto) => {
                     if (!existingPhoto) {
                         let savephote = newInstagramPhotos.save();
-                        let updateuserphoto = UserAllMedia.findOneAndUpdate({ _id: request_object[i]._id }, { sendForReview: true }, { overwrite: true });
-                        Promise.all([savephote,updateuserphoto]).then((response) => {
+                        let updateuserphoto = UserAllMedia.findOneAndUpdate({
+                            _id: request_object[i]._id
+                        }, {
+                            sendForReview: true
+                        }, {
+                            overwrite: true
+                        });
+                        Promise.all([savephote, updateuserphoto]).then((response) => {
                             console.log('added new instagram photo');
                             passed++;
                         }).catch((err) => {
@@ -263,11 +295,13 @@ exports.importUserPhotos = (req, res, next) => {
             }
             if (failed <= 0) {
 
-                newAdminTask(req,res,req.user,request_object);
-                
-            }
-            else {
-                res.status(489).send({ "failed": failed, "passed": passed });
+                newAdminTask(req, res, req.user, request_object);
+
+            } else {
+                res.status(489).send({
+                    "failed": failed,
+                    "passed": passed
+                });
             }
         } else {
             res.status(489).send('Invalid Input');
@@ -278,22 +312,29 @@ exports.importUserPhotos = (req, res, next) => {
 
 }
 
-newAdminTask=(req,res,user,request_object)=>{
-    const adminTask =new AdminTask({user:user,userMedia:request_object});
+newAdminTask = (req, res, user, request_object) => {
+    const adminTask = new AdminTask({
+        user: user,
+        userMedia: request_object
+    });
     adminTask.save((err) => {
         if (err) {
             //return next(err);
             console.log('error while creating admin task');
-            res.status(489).send({ err: "err posting  photos" });
+            res.status(489).send({
+                err: "err posting  photos"
+            });
 
-            
+
         }
         console.log('added new admin task');
-        res.status(200).send({ success: "sucessfully posted  photos" });
+        res.status(200).send({
+            success: "sucessfully posted  photos"
+        });
 
 
     });
-// console.log(user);
+    // console.log(user);
 }
 
 
@@ -304,7 +345,9 @@ newAdminTask=(req,res,user,request_object)=>{
  */
 exports.getAllUnSubmitedPhotos = (req, res, next) => {
 
-    UserAllMedia.find({ facebookId: req.user.facebook }).then((result) => {
+    UserAllMedia.find({
+        facebookId: req.user.facebook
+    }).then((result) => {
         res.status(200).send(result);
 
     }).catch((err) => {
@@ -321,7 +364,9 @@ exports.getAllUnSubmitedPhotos = (req, res, next) => {
  * Unlink OAuth provider.
  */
 exports.getOauthUnlink = (req, res, next) => {
-    const { provider } = req.params;
+    const {
+        provider
+    } = req.params;
     User.findById(req.user.id, (err, user) => {
         if (err) {
             return next(err);
@@ -332,7 +377,9 @@ exports.getOauthUnlink = (req, res, next) => {
             if (err) {
                 return next(err);
             }
-            req.flash('info', { msg: `${provider} account has been unlinked.` });
+            req.flash('info', {
+                msg: `${provider} account has been unlinked.`
+            });
             res.redirect('/account');
         });
     });
@@ -347,14 +394,18 @@ exports.getReset = (req, res, next) => {
         return res.redirect('/');
     }
     User
-        .findOne({ passwordResetToken: req.params.token })
+        .findOne({
+            passwordResetToken: req.params.token
+        })
         .where('passwordResetExpires').gt(Date.now())
         .exec((err, user) => {
             if (err) {
                 return next(err);
             }
             if (!user) {
-                req.flash('errors', { msg: 'Password reset token is invalid or has expired.' });
+                req.flash('errors', {
+                    msg: 'Password reset token is invalid or has expired.'
+                });
                 return res.redirect('/forgot');
             }
             res.render('account/reset', {
@@ -380,25 +431,29 @@ exports.postReset = (req, res, next) => {
 
     const resetPassword = () =>
         User
-            .findOne({ passwordResetToken: req.params.token })
-            .where('passwordResetExpires').gt(Date.now())
-            .then((user) => {
-                if (!user) {
-                    req.flash('errors', { msg: 'Password reset token is invalid or has expired.' });
-                    return res.redirect('back');
-                }
-                user.password = req.body.password;
-                user.passwordResetToken = undefined;
-                user.passwordResetExpires = undefined;
-                return user.save().then(() => new Promise((resolve, reject) => {
-                    req.logIn(user, (err) => {
-                        if (err) {
-                            return reject(err);
-                        }
-                        resolve(user);
-                    });
-                }));
-            });
+        .findOne({
+            passwordResetToken: req.params.token
+        })
+        .where('passwordResetExpires').gt(Date.now())
+        .then((user) => {
+            if (!user) {
+                req.flash('errors', {
+                    msg: 'Password reset token is invalid or has expired.'
+                });
+                return res.redirect('back');
+            }
+            user.password = req.body.password;
+            user.passwordResetToken = undefined;
+            user.passwordResetExpires = undefined;
+            return user.save().then(() => new Promise((resolve, reject) => {
+                req.logIn(user, (err) => {
+                    if (err) {
+                        return reject(err);
+                    }
+                    resolve(user);
+                });
+            }));
+        });
 
     const sendResetPasswordEmail = (user) => {
         if (!user) {
@@ -419,7 +474,9 @@ exports.postReset = (req, res, next) => {
         };
         return transporter.sendMail(mailOptions)
             .then(() => {
-                req.flash('success', { msg: 'Success! Your password has been changed.' });
+                req.flash('success', {
+                    msg: 'Success! Your password has been changed.'
+                });
             });
     };
 
@@ -450,7 +507,9 @@ exports.getForgot = (req, res) => {
  */
 exports.postForgot = (req, res, next) => {
     req.assert('email', 'Please enter a valid email address.').isEmail();
-    req.sanitize('email').normalizeEmail({ gmail_remove_dots: false });
+    req.sanitize('email').normalizeEmail({
+        gmail_remove_dots: false
+    });
 
     const errors = req.validationErrors();
 
@@ -464,17 +523,21 @@ exports.postForgot = (req, res, next) => {
 
     const setRandomToken = token =>
         User
-            .findOne({ email: req.body.email })
-            .then((user) => {
-                if (!user) {
-                    req.flash('errors', { msg: 'Account with that email address does not exist.' });
-                } else {
-                    user.passwordResetToken = token;
-                    user.passwordResetExpires = Date.now() + 3600000; // 1 hour
-                    user = user.save();
-                }
-                return user;
-            });
+        .findOne({
+            email: req.body.email
+        })
+        .then((user) => {
+            if (!user) {
+                req.flash('errors', {
+                    msg: 'Account with that email address does not exist.'
+                });
+            } else {
+                user.passwordResetToken = token;
+                user.passwordResetExpires = Date.now() + 3600000; // 1 hour
+                user = user.save();
+            }
+            return user;
+        });
 
     const sendForgotPasswordEmail = (user) => {
         if (!user) {
@@ -499,7 +562,9 @@ exports.postForgot = (req, res, next) => {
         };
         return transporter.sendMail(mailOptions)
             .then(() => {
-                req.flash('info', { msg: `An e-mail has been sent to ${user.email} with further instructions.` });
+                req.flash('info', {
+                    msg: `An e-mail has been sent to ${user.email} with further instructions.`
+                });
             });
     };
 
