@@ -6,22 +6,10 @@ const User = require('../models/User');
 
 const adminTask = new mongoose.Schema({
 
-    user: {
-        type: {},
-        required: true
-    },
-    userMedia: {
-        type: [],
-        required: true
-    },
-    published: {
-        type: Boolean,
-        required: false,
-        default: false
-    }
-}, {
-    timestamps: true
-});
+    user: {type: {}, required: true},
+    userMedia:{type:[],required:true},
+    published:{type:Boolean,required:false,default:false}
+}, {timestamps: true});
 
 
 /**
@@ -30,5 +18,5 @@ const adminTask = new mongoose.Schema({
 
 
 const AdminTask = mongoose.model('AdminTask', adminTask);
-
+ 
 module.exports = AdminTask;
